Fix migration bug and refactor
