+++ .gitignore
@@ -10,18 +10,7 @@ media
 *.bak 
 
 # If you are using PyCharm # 
-.idea/**/workspace.xml 
-.idea/**/tasks.xml 
-.idea/dictionaries 
-.idea/**/dataSources/ 
-.idea/**/dataSources.ids 
-.idea/**/dataSources.xml 
-.idea/**/dataSources.local.xml 
-.idea/**/sqlDataSources.xml 
-.idea/**/dynamic.xml 
-.idea/**/uiDesigner.xml 
-.idea/**/gradle.xml 
-.idea/**/libraries 
+.idea/
 *.iws /out/ 
 
 # Python # 
@@ -114,8 +103,3 @@ GitHub.sublime-settings
 !.vscode/launch.json 
 !.vscode/extensions.json 
 .history
-
-# Migrations #
-mentorq_backend/migrations/
-mentorq_user/migrations/
-mentorq_main/migrations/
