+++ mentorq_api/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MentorqApiConfig(AppConfig):
+    name = 'mentorq_api'
