--- mentorq_backend/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class MentorqBackendConfig(AppConfig):
-    name = 'mentorq_backend'
