+++ mentorq_main/settings.py
@@ -1,22 +1,7 @@
-"""
-Django settings for mentorq_main project.
-
-Generated by 'django-admin startproject' using Django 2.2.4.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/2.2/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/2.2/ref/settings/
-"""
-
 import os
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-
-# Quick-start development settings - unsuitable for production
-# See https://docs.djangoproject.com/en/2.2/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = 'asdfasdfasdf'
@@ -37,8 +22,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework',
+    'mentorq_api.apps.MentorqApiConfig',
     'mentorq_user.apps.MentorqUserConfig',
-    'mentorq_backend.apps.MentorqBackendConfig',
     'corsheaders',
 ]
 
@@ -76,8 +61,6 @@ TEMPLATES = [
 WSGI_APPLICATION = 'mentorq_main.wsgi.application'
 
 # Database
-# https://docs.djangoproject.com/en/2.2/ref/settings/#databases
-
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.sqlite3',
@@ -91,8 +74,6 @@ AUTHENTICATION_BACKENDS = ["mentorq_user.backends.MentorqUserBackend", "django.c
 # custom Mentorq User model
 AUTH_USER_MODEL = "mentorq_user.MentorqUser"
 # Password validation
-# https://docs.djangoproject.com/en/2.2/ref/settings/#auth-password-validators
-
 AUTH_PASSWORD_VALIDATORS = [
     {
         'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator',
@@ -119,8 +100,6 @@ REST_FRAMEWORK = {
     ),
 }
 # Internationalization
-# https://docs.djangoproject.com/en/2.2/topics/i18n/
-
 LANGUAGE_CODE = 'en-us'
 
 TIME_ZONE = 'UTC'
@@ -132,14 +111,12 @@ USE_L10N = True
 USE_TZ = True
 
 # Static files (CSS, JavaScript, Images)
-# https://docs.djangoproject.com/en/2.2/howto/static-files/
-
 STATIC_URL = '/static/'
 
 # CORS Config
 CORS_ORIGIN_ALLOW_ALL = True
 CORS_ALLOW_CREDENTIALS = False
- 
+
 CORS_ALLOW_METHODS = [
     'DELETE',
     'GET',
