+++ mentorq_main/urls.py
@@ -1,26 +1,11 @@
-"""mentorq_main URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/2.2/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
 from django.urls import path, include
 from lcs_client import set_testing
+
 from mentorq_main.settings import DEV
 
 urlpatterns = [
-    path('', include('mentorq_backend.urls')),
-    path('api/', include('mentorq_backend.api.urls')),
+    path('api/', include('mentorq_api.urls')),
     path('admin/', admin.site.urls),
 ]
 if DEV:
