+++ requirements.txt
@@ -1,15 +1,16 @@
-certifi==2019.9.11
+asgiref==3.2.10
+certifi==2020.6.20
 chardet==3.0.4
-Django==2.2.13
-django-cors-headers==3.2.0
-djangorestframework==3.10.2
-djangorestframework-simplejwt==4.3.0
-idna==2.8
+Django==3.0.8
+django-cors-headers==3.4.0
+djangorestframework==3.11.0
+djangorestframework-simplejwt==4.4.0
+idna==2.10
 lcs-client==1.3
 PyJWT==1.7.1
-python-dateutil==2.8.0
-pytz==2019.2
-requests==2.22.0
-six==1.12.0
-sqlparse==0.3.0
-urllib3==1.25.6
+python-dateutil==2.8.1
+pytz==2020.1
+requests==2.24.0
+six==1.15.0
+sqlparse==0.3.1
+urllib3==1.25.9
